Add missing include file reference

+++ MHD/C_src/CORE_C/kemo_zlib_io_c.c
@@ -66,7 +66,7 @@ void rawseek_go_fwd_f(int *ioffset, int *ierr){
 }
 
 void rawread_32bit_f(int *iflag_swap, int *ilength, char *textbuf, int *lenchara){
-    *lenchara =  fread(textbuf, sizeof(char), *ilength, fp_z);
+    *lenchara = (int) fread(textbuf, sizeof(char), *ilength, fp_z);
     /*
     int i;
     printf("iflag_swap %d\n", *iflag_swap);
